Update to sync mozilla version of SDK with internal 4.0 version

+++ mozilla/directory/c-sdk/ldap/libraries/libldap/extendop.c
@@ -1,19 +1,23 @@
 /* -*- Mode: C++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
  *
- * The contents of this file are subject to the Netscape Public License
- * Version 1.0 (the "NPL"); you may not use this file except in
- * compliance with the NPL.  You may obtain a copy of the NPL at
- * http://www.mozilla.org/NPL/
+ * The contents of this file are subject to the Netscape Public
+ * License Version 1.1 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of
+ * the License at http://www.mozilla.org/NPL/
  *
- * Software distributed under the NPL is distributed on an "AS IS" basis,
- * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the NPL
- * for the specific language governing rights and limitations under the
- * NPL.
+ * Software distributed under the License is distributed on an "AS
+ * IS" basis, WITHOUT WARRANTY OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * rights and limitations under the License.
  *
- * The Initial Developer of this code under the NPL is Netscape
+ * The Original Code is mozilla.org code.
+ *
+ * The Initial Developer of the Original Code is Netscape
  * Communications Corporation.  Portions created by Netscape are
- * Copyright (C) 1998 Netscape Communications Corporation.  All Rights
- * Reserved.
+ * Copyright (C) 1998 Netscape Communications Corporation. All
+ * Rights Reserved.
+ *
+ * Contributor(s): 
  */
 #include "ldap-int.h"
 
@@ -100,7 +104,7 @@ ldap_extended_operation(
 	/* fill it in */
 	if ( ber_printf( ber, "{it{tsto}", msgid, LDAP_REQ_EXTENDED,
 	    LDAP_TAG_EXOP_REQ_OID, exoid, LDAP_TAG_EXOP_REQ_VALUE,
-	    exdata->bv_val, exdata->bv_len ) == -1 ) {
+	    exdata->bv_val, (int)exdata->bv_len /* XXX lossy cast */ ) == -1 ) {
 		rc = LDAP_ENCODING_ERROR;
 		LDAP_SET_LDERRNO( ld, rc, NULL, NULL );
 		ber_free( ber, 1 );
